Moved Avahi DNS SD implementation from network.c to dns_sd_avahi.c.
